Split service_keywords.py (Newtend)
