+++ op_robot_tests/tests_files/brokers/newtend.robot
@@ -2,6 +2,7 @@
 Library  Selenium2Screenshots
 Library  String
 Library  DateTime
+Library  newtend_service.py
 
 *** Variables ***
 ${locator.title}                     xpath=//div[@ng-bind="tender.title"]
@@ -66,7 +67,7 @@ Login
   [Documentation]
   ...      ${ARGUMENTS[0]} ==  username
   ...      ${ARGUMENTS[1]} ==  initial_tender_data
-## Inicialisation
+## Initialisation
   #${prepared_tender_data}=   Add_data_for_GUI_FrontEnds   ${ARGUMENTS[1]}
   ${INITIAL_TENDER_DATA}=  Add_data_for_GUI_FrontEnds  ${INITIAL_TENDER_DATA}
   ${INITIAL_TENDER_DATA}=  Update_data_for_Newtend  ${INITIAL_TENDER_DATA}
