+++ op_robot_tests/tests_files/service_keywords.py
@@ -18,7 +18,6 @@ from .initial_data import (
     test_complaint_reply_data, test_question_answer_data,
     test_question_data, test_tender_data, test_tender_data_multiple_lots
 )
-import calendar
 import os
 import time
 
@@ -220,24 +219,6 @@ def convert_time_to_etender_format(isodate):
 
 def procuringEntity_name(INITIAL_TENDER_DATA):
     INITIAL_TENDER_DATA.data.procuringEntity['name'] = u"Повна назва невідомо чого"
-    return INITIAL_TENDER_DATA
-
-##Newtend
-def newtend_date_picker_index(isodate):
-    now = datetime.today()
-    date_str = '01' + str(now.month) + str(now.year)
-    first_day_of_month = datetime.strptime(date_str, "%d%m%Y")
-    mod = first_day_of_month.isoweekday() - 2
-    iso_dt=parse_date(isodate)
-    last_day_of_month = calendar.monthrange(now.year, now.month)[1] 
-    #LOGGER.log_message(Message("last_day_of_month: {}".format(last_day_of_month), "INFO")) 
-    if now.day>iso_dt.day:
-        mod = calendar.monthrange(now.year, now.month)[1] + mod
-    return mod + iso_dt.day
-
-def Update_data_for_Newtend(INITIAL_TENDER_DATA):
-    #INITIAL_TENDER_DATA.data.items[0].classification['description'] = u"Картонки"
-    INITIAL_TENDER_DATA.data.procuringEntity['name'] = u"openprocurement"
     return INITIAL_TENDER_DATA
 
 def subtract_from_time(date_time,substr_min,substr_sec):
